Setup Firebase dependencies
