+++ app/google-services.json
@@ -1,0 +1,39 @@
+{
+  "project_info": {
+    "project_number": "1073448233091",
+    "project_id": "b07-project-e5893",
+    "storage_bucket": "b07-project-e5893.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:1073448233091:android:e43500678198f36e2913b6",
+        "android_client_info": {
+          "package_name": "com.planner"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "1073448233091-97262crr7rhbvi8uungtn8ipv84tehrk.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyCqQHdwkAASY_RmOHnlzx1DrOjezZdfU4k"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "1073448233091-97262crr7rhbvi8uungtn8ipv84tehrk.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
